+++ tests/test_order_Hexatic.py
@@ -65,8 +65,9 @@ class TestHexatic(unittest.TestCase):
 
     def test_weighted_random(self):
         boxlen = 10
-        N = 500
-        box, points = freud.data.make_random_system(boxlen, N, is2D=True)
+        N = 5000
+        box, points = freud.data.make_random_system(
+            boxlen, N, is2D=True, seed=100)
         voro = freud.locality.Voronoi()
         voro.compute(system=(box, points))
 
@@ -136,6 +137,41 @@ class TestHexatic(unittest.TestCase):
         hop = freud.order.Hexatic(k=4, weighted=True)
         hop.compute(system=(box, points), neighbors=voro.nlist)
         npt.assert_allclose(np.absolute(hop.particle_order), 0., atol=1e-5)
+
+    def test_normalization(self):
+        """Verify normalizations for weighted and unweighted systems."""
+        box = freud.Box.square(L=10)
+        points = np.array([
+            [0.0, 0.0, 0.0],
+            [1.0, 0.31, 0.0],
+            [0.17, 1.0, 0.0],
+            [-1.0, 0.25, 0.0],
+            [0.13, -1.0, 0.0]
+        ])
+        query_point_indices = np.array([0, 0, 0, 0])
+        point_indices = np.array([1, 2, 3, 4])
+        rijs = box.wrap(points[point_indices] -
+                        points[query_point_indices])
+        distances = np.linalg.norm(rijs, axis=-1)
+        thetas = np.arctan2(rijs[:, 1], rijs[:, 0])
+        weights = np.array([1, 0.7, 0.3, 0])
+        nlist = freud.NeighborList.from_arrays(
+            len(points), len(points),
+            query_point_indices, point_indices,
+            distances, weights)
+        for k in range(0, 12):
+            # Unweighted calculation - normalized by number of neighbors
+            psi_k = np.sum(np.exp(thetas * k * 1.0j)) / len(nlist)
+            hop = freud.order.Hexatic(k=k)
+            hop.compute(system=(box, points), neighbors=nlist)
+            npt.assert_allclose(psi_k, hop.particle_order[0], atol=1e-5)
+            # Weighted calculation - normalized by total neighbor weight
+            psi_k_weighted = np.sum(nlist.weights * np.exp(thetas * k * 1.0j))
+            psi_k_weighted /= np.sum(nlist.weights)
+            hop_weighted = freud.order.Hexatic(k=k, weighted=True)
+            hop_weighted.compute(system=(box, points), neighbors=nlist)
+            npt.assert_allclose(
+                psi_k_weighted, hop_weighted.particle_order[0], atol=1e-5)
 
     def test_3d_box(self):
         boxlen = 10