+++ ChangeLog.md
@@ -13,8 +13,13 @@ and this project adheres to
 ### Changed
 * NeighborList raises a `ValueError` instead of a `RuntimeError` if provided invalid constructor arguments.
 
+### Deprecated
+* `freud.order.Translational`
+
 ### Fixed
 * Source distributions now include Cython source files.
+* Hexatic order parameter (unweighted) normalizes by number of neighbors instead of the symmetry order k.
+* Particles with an i-j normal vector of [0, 0, 0] are excluded from 2D Voronoi NeighborList computations for numerical stability reasons.
 
 ## v2.3.0 - 2020-08-03